refactor(daemon): move preserveLockInYaml to provider-safety and fix duplicate import
- Export preserveLockInYaml from provider-safety.ts so tests import the
  actual production function, not a local duplicate
- Remove unused yaml import from misc-routes (moved to provider-safety)
- Merge duplicate mkdtempSync imports in test file
